Refactor out of many-protocol by moving types and functions to many-types
