+++ src/many-client/BUILD.bazel
@@ -4,7 +4,6 @@ load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
 package(default_visibility = [
     "//src/many:__pkg__",
     "//src/many-mock:__pkg__",
-    "//src/many-protocol:__pkg__",
 ])
 
 rust_library(
@@ -21,7 +20,6 @@ rust_library(
         "//src/many-identity",
         "//src/many-identity-dsa",
         "//src/many-modules",
-        "//src/many-protocol",
         "//src/many-server",
         "//src/many-types",
     ],
@@ -44,7 +42,6 @@ rust_library(
         "//src/many-identity:many-identity-for-test",
         "//src/many-identity-dsa:many-identity-dsa-for-test",
         "//src/many-modules:many-modules-for-test",
-        "//src/many-protocol:many-protocol-for-test",
         "//src/many-server:many-server-for-test",
         "//src/many-types:many-types-for-test",
     ],
