+++ src/many-client/Cargo.toml
@@ -26,7 +26,6 @@ many-client-macros = { path = "../many-client-macros", version = "0.1.0" }
 many-identity = { path = "../many-identity", version = "0.1.0" }
 many-identity-dsa = { path = "../many-identity-dsa", version = "0.1.0" }
 many-modules = { path = "../many-modules", version = "0.1.0" }
-many-protocol = { path = "../many-protocol", version = "0.1.0" }
 many-types = { path = "../many-types", version = "0.1.0" }
 minicbor = { version = "0.18.0", features = ["derive", "half", "std"] }
 num-derive = "0.3.3"
