+++ src/many-client/src/client.rs
@@ -10,10 +10,10 @@ use many_identity::verifiers::AnonymousVerifier;
 use many_identity::{verifiers, Identity};
 use many_identity_dsa::CoseKeyVerifier;
 use many_modules::base::Status;
-use many_protocol::{
+use many_server::{Address, ManyError};
+use many_types::{
     encode_cose_sign1_from_request, RequestMessage, RequestMessageBuilder, ResponseMessage,
 };
-use many_server::{Address, ManyError};
 use minicbor::Encode;
 use reqwest::{IntoUrl, Url};
 use std::fmt::{Debug, Formatter};
