+++ src/many-identity-dsa/Cargo.toml
@@ -16,6 +16,7 @@ ed25519-dalek = { version = "1.0.1", optional = true }
 hex = "0.4.3"
 many-error = { path = "../many-error", version = "0.1.0" }
 many-identity = { path = "../many-identity", version = "0.1.0" }
+many-types = { path = "../many-types" }
 minicbor = { version = "0.18.0", optional = true }
 once_cell = "1.10"
 p256 = { version = "0.9.0", features = [ "pem", "ecdsa", "std" ] }
@@ -30,7 +31,6 @@ tracing = "0.1.29"
 
 [dev-dependencies]
 proptest = "1.0.0"
-many-protocol = { path = "../many-protocol", version = "0.1.0" }
 many-identity-dsa = { path = ".", features = [ "default", "ecdsa", "ed25519", "serde", "testing" ] }
 serde_test = "1.0.139"
