+++ src/many-identity-dsa/src/impls/ed25519.rs
@@ -385,8 +385,8 @@ pub mod tests {
     fn sign_and_verify_request() {
         let key = generate_random_ed25519_identity();
         let pubkey = key.public_key();
-        let envelope = many_protocol::encode_cose_sign1_from_request(
-            many_protocol::RequestMessageBuilder::default()
+        let envelope = many_types::encode_cose_sign1_from_request(
+            many_types::RequestMessageBuilder::default()
                 .from(key.address())
                 .method("status".to_string())
                 .data(b"".to_vec())
@@ -396,7 +396,7 @@ pub mod tests {
         )
         .unwrap();
 
-        many_protocol::decode_request_from_cose_sign1(
+        many_types::decode_request_from_cose_sign1(
             &envelope,
             &Ed25519Verifier::from_key(&pubkey).unwrap(),
         )
@@ -407,8 +407,8 @@ pub mod tests {
     fn sign_and_verify_response() {
         let key = generate_random_ed25519_identity();
         let pubkey = key.public_key();
-        let envelope = many_protocol::encode_cose_sign1_from_response(
-            many_protocol::ResponseMessageBuilder::default()
+        let envelope = many_types::encode_cose_sign1_from_response(
+            many_types::ResponseMessageBuilder::default()
                 .from(key.address())
                 .data(Ok(b"".to_vec()))
                 .build()
@@ -417,7 +417,7 @@ pub mod tests {
         )
         .unwrap();
 
-        many_protocol::decode_response_from_cose_sign1(
+        many_types::decode_response_from_cose_sign1(
             &envelope,
             None,
             &Ed25519Verifier::from_key(&pubkey).unwrap(),
