+++ src/many-identity-webauthn/BUILD.bazel
@@ -18,7 +18,7 @@ rust_library(
         "//src/many-error",
         "//src/many-identity",
         "//src/many-identity-dsa",
-        "//src/many-protocol",
+        "//src/many-types",
     ],
 )
 
@@ -38,7 +38,7 @@ rust_library(
         "//src/many-error",
         "//src/many-identity:many-identity-for-test",
         "//src/many-identity-dsa:many-identity-dsa-for-test",
-        "//src/many-protocol:many-protocol-for-test",
+        "//src/many-types:many-types-for-test",
     ],
 )
