+++ src/many-identity-webauthn/Cargo.toml
@@ -11,7 +11,7 @@ coset = "0.3.2"
 many-error = { path = "../many-error" }
 many-identity = { path = "../many-identity" }
 many-identity-dsa = { path = "../many-identity-dsa", features = ["ecdsa"] }
-many-protocol = { path = "../many-protocol" }
+many-types = { path = "../many-types" }
 minicbor = "0.18.0"
 once_cell = "1.13.0"
 serde = "1.0.142"
