+++ src/many-identity/BUILD.bazel
@@ -9,7 +9,6 @@ package(default_visibility = [
     "//src/many-identity-webauthn:__pkg__",
     "//src/many-mock:__pkg__",
     "//src/many-modules:__pkg__",
-    "//src/many-protocol:__pkg__",
     "//src/many-server:__pkg__",
     "//src/many-types:__pkg__",
 ])
