+++ src/many-mock/BUILD.bazel
@@ -19,8 +19,8 @@ rust_library(
         "//src/many-identity-dsa",
         "//src/many-identity-webauthn",
         "//src/many-modules",
-        "//src/many-protocol",
         "//src/many-server",
+        "//src/many-types",
     ],
 )
 
@@ -31,8 +31,8 @@ MANY_MOCK_TEST_DEV_DEPS = [
     "//src/many-identity-dsa:many-identity-dsa-for-test",
     "//src/many-identity-webauthn:many-identity-webauthn-for-test",
     "//src/many-modules:many-modules-for-test",
-    "//src/many-protocol:many-protocol-for-test",
     "//src/many-server:many-server-for-test",
+    "//src/many-types:many-types-for-test",
 ]
 
 rust_library(
