+++ src/many-modules/BUILD.bazel
@@ -5,7 +5,6 @@ package(default_visibility = [
     "//src/many:__pkg__",
     "//src/many-client:__pkg__",
     "//src/many-mock:__pkg__",
-    "//src/many-protocol:__pkg__",
     "//src/many-server:__pkg__",
 ])
 
@@ -22,7 +21,6 @@ rust_library(
     ) + [
         "//src/many-error",
         "//src/many-identity",
-        "//src/many-protocol",
         "//src/many-types",
     ],
 )
@@ -44,7 +42,6 @@ rust_library(
         "//src/many-error",
         "//src/many-identity:many-identity-for-test",
         "//src/many-identity-dsa:many-identity-dsa-for-test",
-        "//src/many-protocol:many-protocol-for-test",
         "//src/many-types:many-types-for-test",
     ],
 )
