+++ src/many-modules/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2021"
 many-error = { path = "../many-error", version = "0.1.0" }
 many-identity = { path = "../many-identity", version = "0.1.0" }
 many-macros = { path = "../many-macros", version = "0.1.0" }
-many-protocol = { path = "../many-protocol", version = "0.1.0" }
 many-types = { path = "../many-types", version = "0.1.0" }
 async-trait = "0.1.56"
 coset = "0.3.2"
