--- src/many-protocol/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "many-protocol"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-many-error = { path = "../many-error", version = "0.1.0" }
-many-identity = { path = "../many-identity", version = "0.1.0" }
-many-types = { path = "../many-types", version = "0.1.0" }
-base64 = "0.13.0"
-coset = "0.3.2"
-derive_builder = "0.10.2"
-hex = "0.4.3"
-minicbor = { version = "0.18.0", features = ["derive", "std"] }
-num-derive = "0.3.3"
-num-traits = "0.2.14"
-num-bigint = "0.4.3"
-serde = "1.0.139"
-serde_json = "1.0.82"
-sha2 = "0.10.2"
-signature = "1.3"
-tracing = "0.1.35"
-url = "2.3.1"
-
-[dev-dependencies]
-once_cell = "1.13.0"
-proptest = "1.0.0"
