+++ src/many-server/BUILD.bazel
@@ -5,7 +5,6 @@ package(default_visibility = [
     "//src/many:__pkg__",
     "//src/many-client:__pkg__",
     "//src/many-mock:__pkg__",
-    "//src/many-protocol:__pkg__",
 ])
 
 rust_library(
@@ -22,7 +21,6 @@ rust_library(
         "//src/many-error",
         "//src/many-identity",
         "//src/many-modules",
-        "//src/many-protocol",
         "//src/many-types",
     ],
 )
@@ -46,7 +44,6 @@ rust_library(
         "//src/many-identity:many-identity-for-test",
         "//src/many-identity-dsa:many-identity-dsa-for-test",
         "//src/many-modules:many-modules-for-test",
-        "//src/many-protocol:many-protocol-for-test",
         "//src/many-types:many-types-for-test",
     ],
 )
