+++ src/many-server/Cargo.toml
@@ -31,7 +31,6 @@ lazy_static = "1.4.0"
 many-error = { path = "../many-error", version = "0.1.0" }
 many-identity = { path = "../many-identity", version = "0.1.0", features = ["coset", "raw"] }
 many-modules = { path = "../many-modules", version = "0.1.0" }
-many-protocol = { path = "../many-protocol", version = "0.1.0" }
 many-types = { path = "../many-types", version = "0.1.0" }
 minicbor = { version = "0.18.0", features = ["derive", "half", "std"] }
 num-bigint = "0.4.3"
