+++ src/many-types/Cargo.toml
@@ -18,6 +18,7 @@ num-traits = "0.2.14"
 num-bigint = "0.4.3"
 proptest = { version = "1.0.0", optional = true }
 serde = "1.0.139"
+url = "2.3.1"
 
 [dev-dependencies]
 cbor-diag = "0.1.11"
