+++ src/many/BUILD.bazel
@@ -18,7 +18,6 @@ rust_binary(
         "//src/many-identity-hsm",
         "//src/many-mock",
         "//src/many-modules",
-        "//src/many-protocol",
         "//src/many-server",
         "//src/many-types",
     ],
