+++ src/many/Cargo.toml
@@ -20,7 +20,6 @@ many-identity-dsa = { path = "../many-identity-dsa", version = "0.1.0", features
 many-identity-hsm = { path = "../many-identity-hsm", version = "0.1.0" }
 many-mock = { path = "../many-mock", version = "0.1.0" }
 many-modules = { path = "../many-modules", version = "0.1.0" }
-many-protocol = { path = "../many-protocol", version = "0.1.0" }
 many-types = { path = "../many-types", version = "0.1.0" }
 many-server = { path = "../many-server", version = "0.1.0" }
 anyhow = "1.0.57"
